+++ WEGO_FE/src/apis/feat5/userpostsApis.ts
@@ -1,5 +1,5 @@
 // import axios from 'axios';
-import { authInstance } from './axiosInstance';
+import { authInstance } from './axiosUserInstance';
 import { AxiosError } from 'axios';
 
 export const userpostsApis = async () => {
